Fixed invisible text effect text bug
I don't know why, i'm not sure how, but this bug beyond human
comprehension has been fixed.

Something to do with either assigning elements as consts (i dont think so)
or not making sure the text is reasserted at start of function and assigned
as it's text content, i dont know why it changes to null but it does

Lesson learned: assume the text may change without any reason, at all
so i must store it and when i need to change i... change it to itself first?

Signed-off-by: n1colassilva <nicolassilva0109@gmail.com>